Tidy up and minor fixes
This commit shuffles files around to organise them better. It creates a new script for specifically removing declarations from frameworks that expired more than 3 years ago. It also fixes some small issues highlighted during the last code review.
